+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './app';
-import { Settting } from './const';
+import App from './components/app';
+import { Settting } from './utils/const';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
