+++ src/pages/favorites.tsx
